feat: api 모듈 구조 세분화
