+++ src/apis/mission.ts
@@ -1,0 +1,3 @@
+const missionApi = {};
+
+export default missionApi;
