+++ src/apis/routine.ts
@@ -1,0 +1,3 @@
+export const routineApi = {};
+
+export default routineApi;
